+++ lib/shopify_cli.rb
@@ -26,6 +26,7 @@ require "cli/kit"
 require "smart_properties"
 require_relative "shopify_cli/version"
 require_relative "shopify_cli/migrator"
+require_relative "shopify_cli/exception_reporter"
 
 # Enable stdout routing. At this point all calls to STDOUT (and STDERR) will go through this class.
 # See https://github.com/Shopify/cli-ui/blob/main/lib/cli/ui/stdout_router.rb for more info
@@ -91,7 +92,7 @@ module ShopifyCLI
   autocall(:ErrorHandler) do
     CLI::Kit::ErrorHandler.new(
       log_file: ShopifyCLI.log_file,
-      exception_reporter: nil,
+      exception_reporter: ->() { ShopifyCLI::ExceptionReporter },
     )
   end