+++ lib/shopify_cli/constants.rb
@@ -9,30 +9,46 @@ module ShopifyCLI
       ANALYTICS_ENABLED = :analytics_enabled
     end
 
+    module Bugsnag
+      API_KEY = "773b0c801eb40c20d8928be5b7c739bd"
+    end
+
+    module Config
+      module Sections
+        module ErrorTracking
+          NAME = "error-tracking"
+          module Fields
+            AUTOMATIC_REPORTING = "automatic-reporting"
+          end
+        end
+      end
+    end
+
     module EnvironmentVariables
+      STACKTRACE = "SHOPIFY_CLI_STACKTRACE"
+
       # When true the CLI points to a local instance of
       # the partners dashboard and identity.
       LOCAL_PARTNERS = "SHOPIFY_APP_CLI_LOCAL_PARTNERS"
 
       # When true the CLI points to a spin instance of spin
       SPIN_PARTNERS = "SHOPIFY_APP_CLI_SPIN_PARTNERS"
-
       SPIN_WORKSPACE = "SPIN_WORKSPACE"
-
       SPIN_NAMESPACE = "SPIN_NAMESPACE"
-
       SPIN_HOST = "SPIN_HOST"
 
-      # Set to true when running tests.
+      # Environments
       TEST = "SHOPIFY_CLI_TEST"
-
-      # Set to true when running tests.
       DEVELOPMENT = "SHOPIFY_CLI_DEVELOPMENT"
     end
 
     module Identity
       CLIENT_ID_DEV = "e5380e02-312a-7408-5718-e07017e9cf52"
       CLIENT_ID = "fbdb2649-e327-4907-8f67-908d24cfd7e3"
+    end
+
+    module Links
+      NEW_ISSUE = "https://github.com/Shopify/shopify-cli/issues/new"
     end
   end
 end